test(presentation): reorganize tests into topic-based subdirectories
Mirror src layout: move adapter tests to adapters/, error tests to
errors/, middleware tests to middleware/. Merge presenter contract tests
into single test_presenter_contract.py with one TestPresenterContract
class. Extract shared test doubles to conftest.py. Rename
test_middleware_protocol.py -> test_middleware.py to match
implementation. Add tests for built-in LoggingMiddleware and
TimingMiddleware.

Enforces 1 test file per implementation, 1 test class per file.
