+++ tests/forging_blocks/presentation/errors/test_error_presenter.py
