+++ tests/forging_blocks/presentation/errors/test_error_view_model.py
@@ -3,7 +3,7 @@
 import pytest
 
 from forging_blocks.presentation import ErrorViewModel
-from forging_blocks.presentation.error_message_model import ErrorMessageModel
+from forging_blocks.presentation.errors.error_message_model import ErrorMessageModel
 
 
 @pytest.mark.unit
